Implimented Menu
Implimented a main menu with three button the first one for start game second for Credit scene and last one for quit game. Second thing implemented is a pause menu with 2 button, one to retry and second to go to main menu. At last I implemented a timer countdown and and gamer over scene, this one has a time text showing how long is left to play and when the time runs out it opens a game over scene with a retry button and main menu button.